requirements1.txt
gradio==4.44.1
requests==2.32.3
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
joblib==1.4.2
numpy==2.1.1
google-generativeai==0.8.3
python-dotenv==1.0.1
scikit-learn==1.5.2